+++ www/js/index.js
@@ -1,0 +1,31 @@
+/**
+ * Club Camper Furgovw's Mobile App
+ * Copyright (C) 2012, Club Camper Furgovw (furgovw.org)
+ * Created by Javier Montes (@mooontes - http://mooontes.com)
+ */
+var app = {
+    initialize: function() {
+        console.log('initialize');
+        this.bindEvents();
+    },
+
+    bindEvents: function() {
+        console.log('bindEvents');
+        document.addEventListener("deviceready", this.onDeviceReady, true);
+    },
+
+    onDeviceReady: function() {
+        console.log('Starting onDeviceReady!');
+       
+        furgovw.deviceready();
+         
+   
+        
+  
+  
+        console.log('welcome!');
+    }
+    
+ 
+
+};